+++ tests/test_graph.py
@@ -712,3 +712,6 @@ def test_graph_ignore():
     dg = DependencyGraph(ignore=datetime)
     with pytest.raises(TypeError):
         dg.resolve(Timer)
+
+
+